Refactored out methods from `Communicator`
The `Communicator` class in network.cpp is actually performing the entire
link state algorithm. Therefore, it has been moved to link_state.cpp and
renamed as `LinkState`. This new class is initialized by providing the id
of the node which it is running on, the initial cost file, and the log
file. Its constructor will create the required objects: Socket, Graph,
HeartbeatsTracker, etc., thereby shielding the user from implementation
details.