Add a mutex in `Graph` to avoid data race
A `graph_mutex` is added into `Graph` class. The mutex is locked whenever
a copy of `Graph::nodes` is made and released after. It is also
locked whenever `Graph::nodes` is being updated`. Furthermore,
`Graph::accept_lsa` will also lock the mutex and only release it once
it's done.

1. In `LinkState::monitor_heartbeats`, the `self_node` is deep copied
first, then modified, and later moved back to the graph.

2. In `LinkState::send_lsa_to_neighbors`, the `self_node` is deep copied
to access it neighbors.

3. In `LinkState::handle_lsa_command`, the `self_node` is deep copied
to access it neighbors.

4. In `RouteTracker::ready_states`, a `graph_snapshot` is created by copy
assign it from the real `graph`.